requirements.txt
pandas
numpy
requests
python-dotenv
matplotlib
plotly
seaborn
futu-api
yfinance
nasdaq-data-link
uuid7
optuna
fastapi
uvicorn[standard]
psycopg[binary]
httpx
argon2-cffi
pyjwt[crypto]
slowapi
pytest
boto3
mkdocs-material
